pricing_service/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
python-dotenv
psycopg2-binary
gunicorn
wtforms